+++ src/LinkedInServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Darvis\ApiLinkedin;
 
+use Darvis\ApiLinkedin\Services\LinkedInImages;
 use Darvis\ApiLinkedin\Services\LinkedInOAuth;
 use Darvis\ApiLinkedin\Services\LinkedInOrganizations;
 use Darvis\ApiLinkedin\Services\LinkedInPublisher;
@@ -17,6 +18,7 @@ class LinkedInServiceProvider extends ServiceProvider
         $this->app->singleton(LinkedInOAuth::class);
         $this->app->singleton(LinkedInPublisher::class);
         $this->app->singleton(LinkedInOrganizations::class);
+        $this->app->singleton(LinkedInImages::class);
         $this->app->singleton(LinkedInManager::class);
         $this->app->alias(LinkedInManager::class, 'linkedin');
     }